Move connection options logic to ConnectOptionsFactory
