+++ app/src/main/java/com/twilio/video/app/data/Preferences.kt
@@ -15,7 +15,6 @@
  */
 package com.twilio.video.app.data
 
-import android.content.SharedPreferences
 import com.twilio.video.BandwidthProfileMode
 import com.twilio.video.OpusCodec
 import com.twilio.video.TrackPriority
@@ -24,6 +23,7 @@ import com.twilio.video.app.BuildConfig
 import com.twilio.video.app.data.api.model.Topology
 
 object Preferences {
+    private const val SERVER_DEFAULT = "Server Default"
     const val EMAIL = "pref_email"
     const val DISPLAY_NAME = "pref_display_name"
     const val ENVIRONMENT = "pref_environment"
@@ -69,26 +69,11 @@ object Preferences {
     const val BANDWIDTH_PROFILE_DOMINANT_SPEAKER_PRIORITY = "pref_bandwidth_profile_dominant_speaker_priority"
     val BANDWIDTH_PROFILE_DOMINANT_SPEAKER_PRIORITY_DEFAULT = TrackPriority.STANDARD.name
     const val BANDWIDTH_PROFILE_TRACK_SWITCH_OFF_MODE = "pref_bandwidth_profile_track_switch_off_mode"
-    const val BANDWIDTH_PROFILE_TRACK_SWITCH_OFF_MODE_DEFAULT = "Server Default"
+    const val BANDWIDTH_PROFILE_TRACK_SWITCH_OFF_MODE_DEFAULT = SERVER_DEFAULT
     const val BANDWIDTH_PROFILE_LOW_TRACK_PRIORITY_RENDER_DIMENSIONS = "pref_bandwidth_profile_low_track_priority_dimensions"
-    const val BANDWIDTH_PROFILE_LOW_TRACK_PRIORITY_RENDER_DIMENSIONS_DEFAULT = "Server Default"
+    const val BANDWIDTH_PROFILE_LOW_TRACK_PRIORITY_RENDER_DIMENSIONS_DEFAULT = SERVER_DEFAULT
     const val BANDWIDTH_PROFILE_STANDARD_TRACK_PRIORITY_RENDER_DIMENSIONS = "pref_bandwidth_profile_standard_track_priority_dimensions"
-    const val BANDWIDTH_PROFILE_STANDARD_TRACK_PRIORITY_RENDER_DIMENSIONS_DEFAULT = "Server Default"
+    const val BANDWIDTH_PROFILE_STANDARD_TRACK_PRIORITY_RENDER_DIMENSIONS_DEFAULT = SERVER_DEFAULT
     const val BANDWIDTH_PROFILE_HIGH_TRACK_PRIORITY_RENDER_DIMENSIONS = "pref_bandwidth_profile_high_track_priority_dimensions"
-    const val BANDWIDTH_PROFILE_HIGH_TRACK_PRIORITY_RENDER_DIMENSIONS_DEFAULT = "Server Default"
-}
-
-/*
- * Utility method that allows getting a shared preference with a default value. The return value
- * type is inferred by the default value type.
- */
-inline fun <reified T> SharedPreferences.get(key: String, defaultValue: T): T {
-    return when (defaultValue) {
-        is Boolean -> getBoolean(key, defaultValue) as T
-        is Float -> getFloat(key, defaultValue) as T
-        is Int -> getInt(key, defaultValue) as T
-        is Long -> getLong(key, defaultValue) as T
-        is String -> getString(key, defaultValue) as T
-        else -> throw IllegalArgumentException("Attempted to get preference with unsupported type")
-    }
+    const val BANDWIDTH_PROFILE_HIGH_TRACK_PRIORITY_RENDER_DIMENSIONS_DEFAULT = SERVER_DEFAULT
 }
