+++ app/src/main/java/com/twilio/video/app/sdk/VideoSdkModule.kt
@@ -17,12 +17,19 @@ import dagger.Provides
 class VideoSdkModule {
 
     @Provides
-    fun providesRoomFactory(
+    fun providesConnectOptionsFactory(
         application: Application,
         sharedPreferences: SharedPreferences,
         tokenService: TokenService
+    ): ConnectOptionsFactory =
+            ConnectOptionsFactory(application, sharedPreferences, tokenService)
+
+    @Provides
+    fun providesRoomFactory(
+        application: Application,
+        connectOptionsFactory: ConnectOptionsFactory
     ): VideoClient =
-            VideoClient(application, sharedPreferences, tokenService)
+            VideoClient(application, connectOptionsFactory)
 
     @Provides
     @ApplicationScope
